% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{calculate_dmu_phi}
\alias{calculate_dmu_phi}
\title{Función que calcula el valor de la transformación (phi) de una observación en el grid.}
\usage{
calculate_dmu_phi(grid, cell)
}
\arguments{
\item{grid}{Objeto SVFGrid sobre el cual operar.}

\item{cell}{Posición de la observación en el grid.}
}
\value{
Una lista que contiene los valores de phi para la celda especificada.
}
\description{
Esta función calcula y retorna el valor de phi para una celda específica del
grid, basado en los datos del grid y la posición de la celda.
}
\examples{
# Crear un conjunto de datos de ejemplo
data <- data.frame(x1 = c(1, 2, 3, 4), x2 = c(5, 6, 7, 8), y1 = c(9, 1, 2, 3))

# Definir listas de inputs, outputs y la cantidad de particiones
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2

# Crear la instancia de la clase SVFGrid y llamar al método create_grid
grid_obj <- SVFGrid(data, inputs, outputs, d)
grid_obj <- create_grid.SVFGrid(grid_obj)

# Ejemplo del calculo de phi para una celda dada
cell <- c(0, 0)
phi_list <- calculate_dmu_phi(grid_obj, cell)
print("Vector phi para la celda dada:")
print(phi_list)
}
