% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{get_estimation.SVF}
\alias{get_estimation.SVF}
\title{Estimacion de una DMU}
\usage{
get_estimation.SVF(svf, dmu)
}
\arguments{
\item{svf}{Objeto de clase SVF.}

\item{dmu}{Observación sobre la que se quiere estimar su valor (vector numérico).}
}
\value{
Vector numérico con las estimaciones de cada output.
}
\description{
Esta función calcula la estimación de los outputs para una DMU específica usando el modelo SVF.
}
