% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svfgrid.R
\name{search_contiguous_cell}
\alias{search_contiguous_cell}
\title{Buscar celdas contiguas en SVFGrid}
\usage{
search_contiguous_cell(cell)
}
\arguments{
\item{cell}{Vector que especifica la posición de la celda en el grid.}
}
\value{
Una lista de celdas contiguas a la especificada.
}
\description{
Esta función identifica y retorna las celdas contiguas a una celda especificada
en el grid. Las celdas contiguas son aquellas que comparten al menos un borde
o punto con la celda especificada.
}
\examples{
# Ejemplo de búsqueda de celda contigua
cell <- c(2, 2)
contiguous_cells <- search_contiguous_cell(cell)
print(paste("Celdas contiguas: (", paste(contiguous_cells, collapse = ", "), ")", sep = ""))
}
