% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvf.R
\name{solve}
\alias{solve}
\title{Resolver el modelo SSVF}
\usage{
solve(svf)
}
\arguments{
\item{svf}{Objeto SSVF con un modelo ya entrenado.}
}
\value{
Una lista conteniendo las soluciones para las variables 'w' y 'xi'.
}
\description{
Esta función extrae las soluciones del modelo SSVF después de que ha sido resuelto.
}
\examples{
# Usar datos de prueba
data <- data.frame(x1 = c(1, 2, 3, 4), x2 = c(5, 6, 7, 8), y1 = c(9, 1, 2, 3))

# Definición de inputs, outputs y otros parámetros
inputs <- c("x1", "x2")
outputs <- c("y1")
d <- 2
C <- 1
eps <- 0
method <- 'SSVF'

# Crear y mostrar el objeto SVF
ssvf <- SSVF(method, inputs, outputs, data, C, eps, d)

trained_svf <- train.SSVF(ssvf)

# Resolver el modelo y mostrar resultados
solution_svf <- solve(trained_svf)
}
