% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf.R
\name{SVF}
\alias{SVF}
\title{Crea una nueva instancia de la clase SVF}
\usage{
SVF(method, inputs, outputs, data, C, eps, d)
}
\arguments{
\item{method}{Método SVF que se quiere utilizar.}

\item{inputs}{Inputs a evaluar en el conjunto de datos.}

\item{outputs}{Outputs a evaluar en el conjunto de datos.}

\item{data}{Conjunto de datos a evaluar.}

\item{C}{Valores del hiperparámetro C del modelo.}

\item{eps}{Valores del hiperparámetro épsilon del modelo.}

\item{d}{Valor del hiperparámetro d del modelo.}
}
\value{
Un objeto de la clase 'SVF'.
}
\description{
Esta función inicializa un objeto SVF con los parámetros proporcionados,
encapsulando un modelo SVF con las especificaciones de entrada y salida,
junto con los hiperparámetros del modelo.
}
\examples{
# Método print para la clase SVF
print.SVF <- function(x) {
  cat("Detalles del modelo SVF:\n")
  cat("----------------------------------\n")
  cat(sprintf("Method: \%s\n", x$method))
  cat(sprintf("Inputs: \%s\n", paste(x$inputs, collapse = ", ")))
  cat(sprintf("Outputs: \%s\n", paste(x$outputs, collapse = ", ")))
  cat(sprintf("Dimensiones de Data: \%d rows, \%d columns\n", nrow(x$data), ncol(x$data)))

  cat("Hyperparámetos:\n")
  cat(sprintf("  C: \%f\n", x$c))
  cat(sprintf("  Epsilon: \%f\n", x$eps))
  cat(sprintf("  d (Número de particiones): \%d\n", x$d))

  cat("Data Preview (solo las primeras columnas):\n")
  if (nrow(x$data) > 0) {
    print(head(x$data))
  } else {
    cat("No hay datos.\n")
  }

  cat("----------------------------------\n")
  invisible(x)
}

data <- data.frame(x1 = c(1, 2, 3, 4), x2 = c(5, 6, 7, 8), y1 = c(9, 1, 2, 3))
inputs <- c("x1", "x2")
outputs <- c("y1")
C <- 1
eps <- 0
d <- 2
method <- 'SVF'

svf <- SVF(method, inputs, outputs, data, C, eps, d)
print(svf)
}
