% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svf_solution.R
\name{SVFPrimalSolution}
\alias{SVFPrimalSolution}
\title{Crea un objeto SVFPrimalSolution}
\usage{
SVFPrimalSolution(w, xi)
}
\arguments{
\item{w}{Valores de los pesos w del modelo ya solucionado.}

\item{xi}{Valores xi del modelo ya solucionado.}
}
\value{
Un objeto de clase SVFPrimalSolution.
}
\description{
Crea un objeto SVFPrimalSolution
}
