% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{transformation}
\alias{transformation}
\title{Transformación de valores en el GRID}
\usage{
transformation(x_i, t_k)
}
\arguments{
\item{x_i}{Valor de la celda a evaluar.}

\item{t_k}{Valor del nodo con el que se quiere comparar.}
}
\value{
Resultado de la comparación: 1, 0, -1.
}
\description{
Esta función evalúa si el valor de una observación es mayor, igual,
o menor que el valor de un nodo en el grid. Devuelve 1 si es mayor,
0 si es igual, y -1 si es menor.
}
\examples{
# Evaluar una observación menor que el valor del nodo del grid
result <- transformation(2, 5)
cat("Resultado de la transformación:", result, "\n")
# Output: -1

# Evaluar una observación igual al valor del nodo del grid
result <- transformation(3, 3)
cat("Resultado de la transformación:", result, "\n")
# Output: 0

# Evaluar una observación mayor que el valor del nodo del grid
result <- transformation(4, 2)
cat("Resultado de la transformación:", result, "\n")
# Output: 1
}
